requirements.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.46.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mysqlclient==2.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
